speed up 9p2
Even though the free list is sorted,
searching via clojure's sequence functions
over and over was slow.  It's a big list.
Searching with a loop dropped part 2 from 3.7s to 440ms.
Still not great, but not wholesale embarassing.

Part1, I went mutable as it's a simple search.
Part2, I went back to idiomatic clojure and the
size of the search space is really making
Clojure's allocation heavy abstractions work hard.